// generated/src/app/api-models/eft-summary-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class EftSummaryWork {
  eftTransNumber: string;
  eftTransDate: Date;
  seqCkprtId: number;
  seqVendId: number;
  seqVendAddress: number;
  vendorId: string;
  fullName: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  vendBankAccountNumber: string;
  vendBankAccountDescription: string;
  vendAbaRoutingNumber: string;
  bankAccountNumber: string;
  bankAccountDescription: string;
  abaRoutingNumber: string;
  bankAccountCode: string;
  netAmt: number;
  amtText: string;
  eftNotes: string;
  eftStatus: string;
  country: string;
  category: string;
}